app/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
drug-smile-fet==1.0.5
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.39.1
MarkupSafe==2.1.2
numba==0.56.4
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
pynndescent==0.5.8
python-dateutil==2.8.2
pytz==2022.7.1
rdkit-pypi==2022.9.5
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
umap-learn==0.5.3
urllib3==1.26.14
Werkzeug==2.2.3
